requirements.txt
django

djangorestframework
djangorestframework-simplejwt

django-cors-headers

dj-rest-auth>=2.2.0
django-allauth>=0.54.0

Pillow

requests